feat: add signing in with WebAuthn
Currently doesn't use the "Conditional UI" feature
of the WebAuthn API, so it's not as convenient as
it could be.

Additionally, the client doesn't display the
registered passkeys, so the user can't delete them.
The client will probably be redone with a new
design, and an upgrade to Svelte 5, so the current
WebAuthn UI looks like a placeholder.